Issues: Fixes #1
Problem:
The issue is that there does not exist any test automation for
continuous integration via Travis.

Analysis:
This initiates the test automation via Travis.  This is a very vanilla
setup with no entrypoint and only minimal items for tox to work with
both commands that are in the .travis_yml file.  This includes the setup
for travis and the basic framework around getting .rpm and .deb
installed via test automation in the framework.

Tests:
There exists a f5_cccl/test/test_f5_cccl.py which is a basic test to
validate that there exists a __version__ variable.