Issues: Fixes #5
Problem:
That the _f5.py and the common.py exists in multiple repos; however, it
is not share across a common repo.  This repo, f5-cccl overlaps on the
functionality of these files; thus, we need to consolidate the location
of these modules

Analysis:
By adding this in a shared module of f5-cccl, we offer a single source
of truth for both of these modules.  These modules will be the starting
point moving forward for velcro on porting over to new work in this
repo.

The change I made was:

* Migrated the modules over into the f5-cccl
* These modules can be imported via `import f5_cccl._f5.py`
 * Or `import f5_cccl.common`

Tests:
There did not exist any unit tests within the two repos that these files
came from; thus, there is no test here yet besides import-ability via
the f5_cccl/__init__.py file and its associated test_f5_cccl.py py.test.