Collect existing BIG-IP config state
Problem:
Need to access the BIG-IP, fetch the state of BIG-IP resources,
and maintain a cache of these resources.

Solution:
Created a CommonBigIP class (very similar to CloudBigIP from
_f5.py) that wraps the f5.bigip class. The refresh() method calls
the get_collection() methods from the SDK, instantiates the resource
objects, and saves them in the cache.

Only implemented for pools and pool members; resource classes for
virtuals, services (iapps), monitors, and policies are not yet
defined.