Modify _f5.py to accept a list of pools
Problem:
Pools and health monitors are currently nested under servers, but need
to be top-level objects.

Solution:
Moved the pools and health monitors out from servers into their own
dicts at the top level. Renamed 'services' to 'virtualServers' and
'policies' to 'l7Policies' to use the same names as cccl. There are
still some differences between this flattening effort and the cccl
schema, but they are beyond the scope of this story.

Fixes VEL-947

affects-branches: master