Allow unattached pools with no related service
Problem: Cloud foundry requires unattached pools which are connected to
a routing virtual server only through policy rules. Kubernetes needs to
support iApps and shouldn't create two pools if a pool is configured to
match the virtual server, but the virtual server is iApp configured. The
restriction is too strong in regards to iApps and prevents cloud foundry
unattached pools from being created.

Solution: Add another condition that allows for unattached pools. If the
pool follows k8s naming conventions it can be found in the f5_services
structure. If the VS found there is configured with an iApp do not add
the pool to the config. If the pool is an unattached orphan and not
found in the f5_services it is ok to configure and won't be a duplicate.

Unit tests and manual tests in CF environment.