Fixes #120:  Handle exceptions when creating iControl REST objects
Problem:
In the bigip common refresh function it is possible for one
of the Icr* constructors to raise an exception. We try to
manage this by handling exceptions for the creation of these
objects and creating a spare ICR object with just a name and
partition. This way the object can be updated or deleted.

Analysis:
Create a new function that will create any iControl REST derived
resource object.

Tests:
test/test_bigip.py