Fixes #122: Reenable auto-created virtual addresses
Problem:
Any virtual address that was created as a result of being defined
as the destination of a virtual server (i.e. auto-created) can be
disabled on the BigIP. A subsequent invocation of apply_config will
not reenable this virtual address because it does not exist in the
configuration.

Analysis:
Add logic to detect when an auto-created virtual address gets
disabled and reenable it.

Tests:
Manual/marathon-bigip