Fixes #128: Modify API to accept a BIG-IP ManagementRoot
Problem:
The initial CCCL API used the BIG-IP credentials to instantiate
a CommonBigIP object which was a sub-class of f5.bigip.ManagmentRoot.

This approach was inflexible in that it requires a session be
established for every managed partition on the BIG-IP.

Analysis:
The API was updated to accept an already instantiated bigip mangement
root.  This change requires modifying CommonBigIP from a sub-class
of ManagementRoot to a client.  The name was changed from CommonBigIP
to BigIPProxy to more accurately reflect its purpose, which is
to serve a proxy/cache for CCCl requests to a BIG-IP.

The change required that the test object be restructured as well.
I modified the TestBigIP object to a mock BigIP.

Tests:
f5_cccl/test/test_bigip.py
f5_cccl/service/test/test_service_manager.py