Allow setting caFile in server SSL profiles
Problem:
When peerCert is set to "require" in the server SSL profile it uses the
caFile member (Trusted Certificate Authorities in the Server
Authentication section as viewed in the UI) to validate the server
certificate, not the Chain member in Configuration section.

Solution:
Added knowledge of the caFile member in SSL profiles. The default value
of caFile is an empty string which is consistent with previous
behavior, but it can also be the path to the CA certificate on the
BIG-IP (/Common/some-ca.crt) or 'self' which indicates that the profile
contains the CA certificate itself and that a SSL profile should NOT be
configured, just install the certificate.

affects-branches: master