Fix tunnel record update logic
Problem: Records were not being sorted as expected, and resulted in tunnels being updated
unnecessarily due to records being equal in value but not order.

Solution: Rather than relying on sorting records (which we don't really need to do), change
the comparison logic for records to prevent updates that are not needed.