Set "source" default for Virtual resource
Problem: The CF controller uses the "source" field on virtuals, but
the k8s controller does not. This field was processed as a default field,
but since the k8s controller did not populate the field, there were constant
virtual server updates due to mismatching fields.

Solution: Set a source default if one is not provided, preventing any unnecessary
updates if the field is not provided by a controller.