Enforce relationship between health monitor interval and timeout
Problem: For Health Monitors to function correctly, the interval
value should be less than the timeout value. BIG-IP versions prior
to v13.0 did not enforce this; however BIG-IP v13.0 does. Also,
BIG-IP v13.0 will actually change the interval value to be less than
the timeout if the input interval is greater than the timeout.

Solution: Enforce this relationship for API Health Monitors in CCCL.
Ignore it for ICR Health Monitors because these represent objects
that already exist on BIG-IP that we can't ignore.