Support whitelisted resources
Problem: To have feature parity with the original OpenShift
Big-IP router, we need to allow the user the ability to create
resources and then have the controller manage them.

Solution: CCCL will keep track of changes made by the CCs to
whitelisted resources. These are resources that have a
cccl-whitelist metadata field set. The changes will be kept
track of in another metadata field named cccl-whitelist-updates.
When new changes are requested by the CC, the old changes will
be backed out and the new ones added. If the resource is no longer
being controlled, then the old changes will simply be backed out.