Pin f5-icontrol-rest
Problem: The f5-common-python which is pinned by the CCCL, does not
pin the f5-icontrol-rest package this it imports. This introduces
the possibility to pull in unwanted changes and makes reproducible
builds impossible.

Solution: The temporary solution is to have the CCCL pin the
f5-icontrol-rest to a known good version.