Prune profiles of whitelisted virtuals
Problem:  The handling of unused profiles is done out-of-band.
That is, outside of the call apply_ltm_config().  This results
in REST errors when CCCL tries to remove profiles that are in
used by white-listed virtuals.

Solution: Similar to the handling in apply_ltm_config, a list of
profiles is create that are referenced from whitelisted virtuals.
When the request to delete unused profiles is made (based on the
config), this list is consulted and if a match is found, the
request to delete is ignored.

Also removed a duplicate function.

Fixes: #220, #223