Force update if user deletes the cccl-update flag
Problem: If a resource transitions from being unmanaged (whitelisted)
to managed, we may end up leaving the cccl-whitelist-updates metadata
key (which contains the latest patch). This occurs only if the
cccl-whitelist metadata key is deleted instead of being set to false.

Solution: Force the cccl-whitelist metadata field to be set to false if we
detect the condition mentioned above.  This will in turn force the resource
to be updated (since the desired config of a managed resource doesn't specify
the cccl-whitelist metadata field).  The update itself will permenantly clear
both the metadata fields.

Fixes: #226